+++ extensions/flabel/src/arc/flabel/effects/WindEffect.java
@@ -34,8 +34,8 @@ public class WindEffect extends FEffect{
 
         // Calculate noise
         float indexOffset = localIndex * 0.05f * spacing;
-        float noiseX = (float) Simplex.noise2d(1, 6, 0, 1f, noiseCursorX + indexOffset, 0);
-        float noiseY = (float) Simplex.noise2d(1, 6, 0, 1f, noiseCursorY + indexOffset, 0);
+        float noiseX = Simplex.noise2d(1, 6, 0, 1f, noiseCursorX + indexOffset, 0);
+        float noiseY = Simplex.noise2d(1, 6, 0, 1f, noiseCursorY + indexOffset, 0);
 
         // Calculate offset
         float lineHeight = getLineHeight(label);
